requirements/base.txt
asgiref==3.5.2
autopep8==2.0.0
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
crispy-bootstrap5==0.2
dj-database-url==0.5.0
Django==3.2.16
django-appconf==1.0.4
django-choices==1.7.2
django-compressor==2.4
django-cors-headers==3.5.0
django-crispy-forms==1.11.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-libsass==0.8
django-ses==2.0.0
django-storages==1.11.1
django-tinymce4-lite==1.8.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.21.4
durationpy==0.5
future==0.18.2
gunicorn==20.0.4
hjson==3.0.2
idna==2.10
install==1.3.4
jmespath==0.10.0
jsmin==3.0.1
kappa==0.6.0
libsass==0.20.1
Markdown
MarkupSafe==1.1.1
mccabe==0.7.0
openpyxl==3.1.2
pdfkit==1.0.0
pep517==0.10.0
pip-tools==6.0.1
pillow==9.4.0
placebo==0.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
PyYAML==5.4.1
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.59.0
typing_extensions==4.4.0
urllib3==1.26.4
Werkzeug==0.16.1
whitenoise==5.2.0
wsgi-request-logger==0.4.6
zipp==3.3.2
